Set sync and override to allow overwriting
The logic within the Skin package does not make sense, and the code
base is borderline unreadable due to different entry-points and
controller actions being merged in to single monolith class rather than
using dependency injection, or breaking collections and entities apart;
it is actually half singleton and half instance based.

No one looking at the class have zero idea what they should be calling,
and majority of the documentation and hinting is invalid, and it would
be rather easy to call the wrong method.